I tried following Pascal's ERRest presentations at WOWODC 2012 and running this code at the same time. I found that, because Pascal was changing the code during the presentation, demonstrations at the beginning would no longer work at the end. These changes should bring the code more in line with the presentations.
1. Would not even run. I have fixed a null pointer exception on startup
caused by leaving the method null when inserting the BlogEntry route for
"/" in Application constructor.

2. Some of the methods Pascal added during the presentation did not make
their way into committed code, such as Authors destroyAction(),
indexAction(), showAction(), updateAction(). I have added them now.

3. Added some comments

4. Fixed BlogEntryShowPage so that it works with the new uniqueTitle id
for Blog Entries which is implemented in the final code.

5. Added a help screen with instructions for how to test the program
using curl on the command line. (Main.wo) This saved me a lot of typing
into HTTP Client when I was testing. It also shows the final version of
the URLs for those wondering why the URLs Pascal uses early in
presentation are not working.

6. I changed the cookie domain for the login action so that the cookie
generated during login could be used unchanged in the BlogEntries
actions. The curl scripts are designed to store the cookie in
/tmp/BlogDemoCookieJar.txt from login and use cookies from that file
later when working with BlogEntries.

7. Changed the BlogEntryRestDelegate to use the full REST id (title and
numeric eof id) of the BlogEntry and not just the title bit